Extract post edit request construction
